# app/backend/api.yml
...
openapi: 3.0.0
info:
  title: CHEF Backend API
  version: 1.0.0
...
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /recipe/{recipeId}/similar:
    get:
      tags: [Recipe]
      summary: Get similar recipes
      description: Returns a list of recipes similar to the given recipe \
        Items are sorted by similarity score, descending \
        Only recipes of the same type are returned
        Note that if multiple recipes have the same name, only one will be returned
      parameters:
        - $ref: "#/components/parameters/recipeId"
        - $ref: "#/components/parameters/limitRequired"
        - $ref: "#/components/parameters/minSimilarity"
        # TODO: Authentication if filtering by fridge
        - in: query
          name: availableForFridge
          description: If specified, only return recipes that can be made with the ingredients in the fridge
          schema: { type: integer, example: 1 }
          required: false

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimilarRecipe"
        "404":
          $ref: "#/components/responses/NotFound"

  # TODO: Authentication for all fridge endpoints
  # TODO: Implement
...
  /fridge/{fridgeId}/recipe/available:
    get:
      tags: [Fridge]
      summary: Get a list of available recipes
      description: Returns the IDs of all recipes that can be made with the available ingredients
      parameters:
        - $ref: "#/components/parameters/fridgeId"
        - in: query
          name: maxMissingIngredients
          description: Maximum number of ingredients that can be missing.
          schema: { type: integer, minimum: 0, example: 1, default: 0 }
          required: false
        - in: query
          name: checkAmounts
          description: Whether to check that there is enough of each ingredient.
          schema: { type: boolean, example: true, default: true }
          required: false
        - in: query
          name: mealType
          description: If specified, only return recipes of this type
          schema: { type: string, example: "Dinner" }
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name: { type: string, example: Just Soup }
                    id: { type: integer, example: 1234 }
                    missingIngredientAmount:
                      {
                        type: integer,
                        example: 1,
                        description: Number of ingredients missing or not enough of,
                      }
                  required: [name, id, missingIngredientAmount]
        "403":
          $ref: "#/components/responses/Forbidden"

  # TODO: Authentication
...
components:
  parameters:
    fridgeId:
      in: path
      name: fridgeId
      schema: { type: integer, example: 1 }
      required: true
...
    recipeId:
      in: path
      name: recipeId
      schema: { type: integer, example: 1 }
      required: true
...
    limitRequired:
      in: query
      name: limit
      schema: { type: integer, example: 10 }
      required: true
    minSimilarity:
      in: query
      name: minSimilarity
      schema: { type: number, example: 0.5, minimum: 0, maximum: 1, default: 0.5 }
      required: false
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorList" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorList" }
...
  schemas:
    ErrorList:
      type: object
      properties:
        status: { type: integer, example: 404 }
        path: { type: string, example: "/api/v1/path/to/endpoint" }
        errors:
          type: array
          items: { $ref: "#/components/schemas/Error" }
      required: [status, path, errors]
    Error:
      type: object
      properties:
        path: { type: string, example: "/params/id" }
        message: { type: string, example: "Must be integer" }
        errorCode: { type: string, example: "type.openapi.validation" }
      required: [path, message]
...
    SimilarRecipe:
      type: object
      properties:
        name: { type: string, example: Just Soup }
        id: { type: integer, example: 1234 }
        similarity: { type: number, example: 0.5 }
      required: [name, id, similarity]
...